Telephony MSIM: Change subid from long to int.
Change subid from long to int for getActiveSubscription()
and switchToOtherActiveSub() APIs.

CRs-Fixed: 810354

+++ telecomm/java/android/telecom/TelecomManager.java
@@ -931,7 +931,7 @@ public class TelecomManager {
      * Returns current active subscription.
      * @hide
      */
-    public long getActiveSubscription() {
+    public int getActiveSubscription() {
         try {
             if (isServiceConnected()) {
                 return getTelecomService().getActiveSubscription();
@@ -946,7 +946,7 @@ public class TelecomManager {
      * switches to other active subscription.
      * @hide
      */
-    public void switchToOtherActiveSub(long subId) {
+    public void switchToOtherActiveSub(int subId) {
         try {
             if (isServiceConnected()) {
                 getTelecomService().switchToOtherActiveSub(subId);
